Rearrange code so that SLO stuff is in slo.py and Slack stuff is in slack.py. Fix a bug where SLOs failing/recovering during the abandon countdown would cancel the abandon countdown.
