+++ paasta_tools/automatic_rollbacks/state_machine.py
@@ -82,6 +82,11 @@ class DeploymentProcess(abc.ABC):
     def start_state(self):
         raise NotImplementedError()
 
+    @abc.abstractmethod
+    def notify_users(self, message: str) -> None:
+        """Print a log line somewhere that users can see it, e.g. Slack."""
+        raise NotImplementedError()
+
     def is_terminal_state(self, state: str) -> bool:
         return (state in self.status_code_by_state())
 
@@ -128,13 +133,17 @@ class DeploymentProcess(abc.ABC):
 
     def cancel_timer(self, trigger=None):
         """Cancel the running timer. If trigger is specified, only cancel the timer if its trigger matches."""
-        self.timer_running = False
         handle = self.get_timer_handle()
         if handle is None:
+            self.timer_running = False
             return
         if trigger is None or trigger == self.timer_trigger:
+            self.notify_users(f"Countdown to {self.timer_message_verb} ({self.timer_trigger} == {trigger}) cancelled.")
             handle.cancel()
             self.timer_handle = None
+            self.timer_running = False
+            self.timer_trigger = None
+            self.timer_message_verb = None
 
     def restart_timer(self):
         self.cancel_timer()
@@ -154,4 +163,4 @@ class DeploymentProcess(abc.ABC):
             return None
 
     def before_state_change(self):
-        self.cancel_timer()
+        pass
